requirements.txt
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
et-xmlfile==1.0.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
numpy==1.20.1
openpyxl==3.0.7
pandas==1.2.3
parsel==1.6.0
Pebble==4.6.1
Protego==0.1.16
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
queuelib==1.5.0
requests==2.25.1
Scrapy==2.4.1
selenium==3.141.0
service-identity==18.1.0
site-map-parser==0.3.6
six==1.15.0
soupsieve==2.2.1
Twisted==21.2.0
ultimate-sitemap-parser==0.5
url-normalize==1.4.3
urllib3==1.26.4
w3lib==1.22.0
xlrd==2.0.1
zope.interface==5.3.0
